Refactoring: new project's structure
