+++ samples/structured_data/insert_sec_stream_data.py
@@ -1,7 +1,7 @@
-import st_library.dataprovider.structured_data as stdata
-import st_library.dataprovider as datapv
+import st_library.core.dataprovider.structured_data as stdata
+from st_library import Library
 
-st_lib = datapv.Library()
+st_lib = Library()
 st_lib.set_token('Bearer token')
 st_lib.set_configuration_uuid('98a599e9-e5b8-49fc-b5ca-0750bb0a785b')
