+++ samples/structured_data/read_sec_batch_matrice.py
@@ -10,8 +10,9 @@
 # or implied. See the License for the specific language governing permissions and limitations under
 # the License.
 
-import st_library.dataprovider as datapv
-st_lib = datapv.Library()
+from st_library import Library
+
+st_lib = Library()
 st_lib.set_token('Bearer token')
 st_lib.set_configuration_uuid('98a599e9-e5b8-49fc-b5ca-0750bb0a785b')
