+++ st_library/core/dataprovider/structured_data/_schema.py
