+++ st_library/core/dataprovider/unstructured_data/_api.py
@@ -14,7 +14,7 @@
 
 import sys
 
-from st_library.dataprovider import _http
+from st_library.utils.api_client import _http
 
 
 class Api(object):
