+++ st_library/utils/generics/singleton.py
